restructuring of users API
