+++ tests/io/v3/test_schema.py
@@ -3,7 +3,7 @@ Testing the users schemas
 '''
 import pytest
 from marshmallow.exceptions import ValidationError
-from tenable.io.v3.vm.schema import (
+from tenable.io.v3.schema import (
     UserEditSchema,
     UsersCreateSchema
 )
